Add PostList and Code Cleanup
