+++ src/MainContainer/components/AppBar.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LogoComponent from "./LogoComponent";
+import { ReactComponent as MainLogo } from "../images/main-logo.svg";
 import "../css/AppBar.css";
 
 function AppBar() {
@@ -7,7 +7,7 @@ function AppBar() {
     <header className="appbar">
       <div className="header-container">
         <h1>YAJA</h1>
-        <LogoComponent />
+        <MainLogo />
         <div className="search-form">
           <input type="text" placeholder="search.." />
           <div className="search_logo"></div>
