--- src/modules/index.jsx
@@ -1,6 +1,0 @@
-// import { combineReducers } from "redux";
-// import posts from "./posts";
-
-// const rootReducer = combineReducers({ posts });
-
-// export default rootReducer;
